requirements.txt
aiohttp==3.2.1
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.6.3
astropy==3.0.2
async-timeout==3.0.0
attrs==18.1.0
Automat==0.7.0
autopep8==1.3.2
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
blinker==1.4
bokeh==0.12.15
boto==2.48.0
boto3==1.7.21
botocore==1.10.21
Bottleneck==1.2.1
bz2file==0.98
CacheControl==0.12.4
certifi==2018.8.24
cffi==1.11.5
cfscrape==1.9.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.5.11
constantly==15.1.0
contextlib2==0.5.5
coverage==4.5.1
cryptography==2.2.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.4
datashape==0.5.4
decorator==4.3.0
distlib==0.2.7
distributed==1.21.8
Django==2.0.6
django-haystack==2.8.1
djangorestframework==3.8.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
gensim==3.4.0
gevent==1.3a1
glob2==0.5
gmail3==0.0.6
greenlet==0.4.13
h5py==2.8.0
heapdict==1.0.0
html-purifier==0.1.9
html5lib==1.0.1
hyperlink==17.3.1
hypothesis==3.56.9
idna==2.6
idna-ssl==1.0.0
imageio==2.3.0
imagesize==1.0.0
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.22.0
locket==0.2.0
lockfile==0.12.2
logzero==1.5.0
lxml==4.2.1
lyricwikia==0.1.9
manga-py==1.3.17
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.12
mistune==0.8.3
mkl-fft==1.0.2
mkl-random==1.0.1
more-itertools==4.1.0
msgpack==0.5.6
msgpack-python==0.5.6
multidict==4.2.0
multipledispatch==0.5.0
mutagen==1.40.0
navigator-updater==0.2.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.5.0
numba==0.37.0
numexpr==2.6.5
numpy==1.14.2
numpydoc==0.8.0
oauthlib==2.0.7
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.6
packaging==17.1
pafy==0.5.4
pandas==0.23.0
pandocfilters==1.4.2
parsel==1.4.0
parso==0.2.0
partd==0.3.8
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.2
patsy==0.5.0
pdfrw==0.4
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
progress==1.3
progressbar2==3.38.0
prompt-toolkit==1.0.15
psutil==5.4.7
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAutoGUI==0.9.38
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
PyDispatcher==2.0.5
PyExecJS==1.5.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.4
Pympler==0.5
PyMsgBox==1.0.6
pyodbc==4.0.23
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyreadline==2.1
PyScreeze==0.1.18
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.2.1
pytest-doctestplus==0.1.2
pytest-openfiles==0.2.0
pytest-remotedata==0.2.0
pytest-runner==4.2
python-crfsuite==0.9.2
python-dateutil==2.7.3
python-docx==0.8.6
python-utils==2.3.0
PyTweening==1.0.3
pytz==2018.4
PyWavelets==0.5.2
pywin32==224
pywinpty==0.5
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.2
queuelib==1.5.0
requests==2.19.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rope==0.10.7
rsa==3.4.2
ruamel-yaml==0.15.37
s3transfer==0.1.13
schedule==0.5.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.5.1
scribd-downloader==0.1.5
seaborn==0.8.1
selenium==3.14.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.7
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spotdl==1.0.0
spotipy==2.4.4
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
tika==1.18
titlecase==0.12.0
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
twilio==6.18.1
Twisted==17.5.0
twython==3.7.0
typing==3.6.4
unicode-slugify==0.1.3
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
virtualenv==15.1.0
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
yarl==1.2.4
youtube-dl==2018.4.3
zict==0.1.3
zope.interface==4.5.0
